backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
supabase==2.10.0
langchain==0.3.14
langchain-groq==0.2.3
langgraph==0.2.60
apscheduler==3.10.4
pydantic==2.10.4
python-multipart==0.0.20
aiofiles==24.1.0
websockets==14.1
ffmpeg-python==0.2.0
